rfdetr-object-detection/requirements.txt
litserve
onnxruntime
pillow
python-multipart
rfdetr
supervision
